schema_master: fix pre-existing package nits (VEC-492)
Follow-up nits from Tore's #545 review (pre-existing, not introduced by #545):
- Transform.GuardMin/GuardMax *int -> *int64: guard_max 4100000000 exceeds the
  32-bit int ceiling, so Load() would fail to unmarshal on a 32-bit build.
- Drop the dead tables.migrations entry: migrations is in ignore_tables, which
  every Check pass short-circuits on, so the tables entry was unreachable.
- Document the intentionally-untyped maple_* tables in the README: they are
  skipped by the required-key pass (gated on type), pending classification.
- Remove the redundant non-nullable t.build_id fixture row in TestNullability
  (a real table can't have both; the assertion is unaffected either way).
- Give the conformance integration test a 30s timeout context rather than
  context.Background().

+++ stl-verify/data_quality/schemamaster/schemamaster.go
@@ -45,8 +45,8 @@ type Transform struct {
 	Canonical string `json:"canonical"`
 	Action    string `json:"action"`
 	From      string `json:"from"`
-	GuardMin  *int   `json:"guard_min"`
-	GuardMax  *int   `json:"guard_max"`
+	GuardMin  *int64 `json:"guard_min"`
+	GuardMax  *int64 `json:"guard_max"`
 }
 
 // Override is a sanctioned TYPE exemption: a column deliberately kept at accepted_type rather
